Upadate & correctio des  bugn
