+++ apps/app/src/app/[lang]/dashboard/[section]/page.tsx
@@ -1,8 +1,7 @@
 export default function SectionPage() {
-    return (
-      <div>
-        <h1>Page dynamique pour [section]</h1>
-      </div>
-    );
-  }
-  
+  return (
+    <div>
+      <h1>Page dynamique pour [section]</h1>
+    </div>
+  )
+}
