+++ apps/app/src/app/lib/dossierStore.ts
@@ -1,5 +1,5 @@
 // app/lib/dossierStore.ts
-import{ Dossier } from "@/app/models/Dossier"
+import { Dossier } from "@/app/models/Dossier"
 
 // Simulation d'une base de données (à remplacer par votre vraie DB)
 export const dossiers: Dossier[] = []
