+++ apps/app/src/app/models/Dossier.ts
