+++ apps/app/src/components/dashboard/Card.tsx
@@ -1,6 +1,6 @@
 "use client "
 
-import React from 'react'
+import React from "react"
 
 interface CardProps {
   title: string
@@ -11,7 +11,7 @@ interface CardProps {
 
 const Card: React.FC<CardProps> = ({ title, value, description, icon }) => {
   return (
-    <div className="bg-white p-6 shadow-lg rounded-lg w-full md:w-1/3 lg:w-1/4">
+    <div className="w-full rounded-lg bg-white p-6 shadow-lg md:w-1/3 lg:w-1/4">
       <div className="flex items-center space-x-4">
         <div className="text-4xl text-green-600">{icon}</div>
         <div>
